feat: add on_block
